CLOUD-46412 move template related classes to template modul
