+++ cloud-openstack/src/main/java/com/sequenceiq/cloudbreak/cloud/openstack/nativ/context/OpenStackContext.java
@@ -2,7 +2,7 @@ package com.sequenceiq.cloudbreak.cloud.openstack.nativ.context;
 
 import org.openstack4j.api.OSClient;
 
-import com.sequenceiq.cloudbreak.cloud.event.context.ResourceBuilderContext;
+import com.sequenceiq.cloudbreak.cloud.template.context.ResourceBuilderContext;
 
 public class OpenStackContext  extends ResourceBuilderContext {
     private OSClient osClient;
