+++ cloud-reactor/src/main/java/com/sequenceiq/cloudbreak/cloud/task/PollTaskFactory.java
@@ -11,17 +11,12 @@ import org.springframework.stereotype.Component;
 import com.sequenceiq.cloudbreak.cloud.CloudConnector;
 import com.sequenceiq.cloudbreak.cloud.InstanceConnector;
 import com.sequenceiq.cloudbreak.cloud.event.context.AuthenticatedContext;
-import com.sequenceiq.cloudbreak.cloud.event.context.ResourceBuilderContext;
 import com.sequenceiq.cloudbreak.cloud.event.instance.InstanceConsoleOutputResult;
 import com.sequenceiq.cloudbreak.cloud.event.instance.InstancesStatusResult;
 import com.sequenceiq.cloudbreak.cloud.init.CloudPlatformConnectors;
 import com.sequenceiq.cloudbreak.cloud.model.CloudInstance;
 import com.sequenceiq.cloudbreak.cloud.model.CloudResource;
-import com.sequenceiq.cloudbreak.cloud.model.CloudResourceStatus;
-import com.sequenceiq.cloudbreak.cloud.model.CloudVmInstanceStatus;
 import com.sequenceiq.cloudbreak.cloud.model.InstanceStatus;
-import com.sequenceiq.cloudbreak.cloud.template.ComputeResourceBuilder;
-import com.sequenceiq.cloudbreak.cloud.template.ResourceChecker;
 
 @Component
 public class PollTaskFactory {
@@ -47,15 +42,7 @@ public class PollTaskFactory {
         return createPollTask(PollInstanceConsoleOutputTask.NAME, instanceConnector, authenticatedContext, instance);
     }
 
-    public PollTask<List<CloudResourceStatus>> newPollResourceTask(ResourceChecker checker, AuthenticatedContext authenticatedContext,
-            List<CloudResource> cloudResource, ResourceBuilderContext context, boolean cancellable) {
-        return createPollTask(PollResourceTask.NAME, authenticatedContext, checker, cloudResource, context, cancellable);
-    }
 
-    public PollTask<List<CloudVmInstanceStatus>> newPollComputeStatusTask(ComputeResourceBuilder builder, AuthenticatedContext authenticatedContext,
-            ResourceBuilderContext context, CloudInstance instance) {
-        return createPollTask(PollComputeStatusTask.NAME, authenticatedContext, builder, context, instance);
-    }
 
     @SuppressWarnings("unchecked")
     private <T> T createPollTask(String name, Object... args) {
