+++ cloud-template/src/main/java/com/sequenceiq/cloudbreak/cloud/template/ResourceChecker.java
@@ -3,7 +3,7 @@ package com.sequenceiq.cloudbreak.cloud.template;
 import java.util.List;
 
 import com.sequenceiq.cloudbreak.cloud.event.context.AuthenticatedContext;
-import com.sequenceiq.cloudbreak.cloud.event.context.ResourceBuilderContext;
+import com.sequenceiq.cloudbreak.cloud.template.context.ResourceBuilderContext;
 import com.sequenceiq.cloudbreak.cloud.model.CloudResource;
 import com.sequenceiq.cloudbreak.cloud.model.CloudResourceStatus;
