+++ cloud-template/src/main/java/com/sequenceiq/cloudbreak/cloud/template/compute/ResourceDeleteThread.java
@@ -13,15 +13,15 @@ import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
 import com.sequenceiq.cloudbreak.cloud.event.context.AuthenticatedContext;
-import com.sequenceiq.cloudbreak.cloud.event.context.ResourceBuilderContext;
+import com.sequenceiq.cloudbreak.cloud.template.context.ResourceBuilderContext;
 import com.sequenceiq.cloudbreak.cloud.model.CloudResource;
 import com.sequenceiq.cloudbreak.cloud.model.CloudResourceStatus;
 import com.sequenceiq.cloudbreak.cloud.model.ResourceStatus;
 import com.sequenceiq.cloudbreak.cloud.notification.PersistenceNotifier;
 import com.sequenceiq.cloudbreak.cloud.scheduler.SyncPollingScheduler;
 import com.sequenceiq.cloudbreak.cloud.task.PollTask;
-import com.sequenceiq.cloudbreak.cloud.task.PollTaskFactory;
 import com.sequenceiq.cloudbreak.cloud.template.ComputeResourceBuilder;
+import com.sequenceiq.cloudbreak.cloud.template.task.ResourcePollTaskFactory;
 import com.sequenceiq.cloudbreak.common.type.CommonStatus;
 
 @Component(ResourceDeleteThread.NAME)
@@ -34,7 +34,7 @@ public class ResourceDeleteThread implements Callable<ResourceRequestResult<List
     @Inject
     private SyncPollingScheduler<List<CloudResourceStatus>> syncPollingScheduler;
     @Inject
-    private PollTaskFactory statusCheckFactory;
+    private ResourcePollTaskFactory resourcePollTaskFactory;
     @Inject
     private PersistenceNotifier resourceNotifier;
 
@@ -59,7 +59,7 @@ public class ResourceDeleteThread implements Callable<ResourceRequestResult<List
         if (resource.getStatus() == CommonStatus.CREATED) {
             CloudResource deletedResource = builder.delete(context, auth, resource);
             if (deletedResource != null) {
-                PollTask<List<CloudResourceStatus>> task = statusCheckFactory
+                PollTask<List<CloudResourceStatus>> task = resourcePollTaskFactory
                         .newPollResourceTask(builder, auth, asList(deletedResource), context, cancellable);
                 List<CloudResourceStatus> pollerResult = syncPollingScheduler.schedule(task);
                 deleteResource();
