+++ cloud-template/src/main/java/com/sequenceiq/cloudbreak/cloud/template/task/PollResourceTask.java
@@ -1,4 +1,4 @@
-package com.sequenceiq.cloudbreak.cloud.task;
+package com.sequenceiq.cloudbreak.cloud.template.task;
 
 import java.util.List;
 
@@ -6,9 +6,10 @@ import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
 import com.sequenceiq.cloudbreak.cloud.event.context.AuthenticatedContext;
-import com.sequenceiq.cloudbreak.cloud.event.context.ResourceBuilderContext;
+import com.sequenceiq.cloudbreak.cloud.template.context.ResourceBuilderContext;
 import com.sequenceiq.cloudbreak.cloud.model.CloudResource;
 import com.sequenceiq.cloudbreak.cloud.model.CloudResourceStatus;
+import com.sequenceiq.cloudbreak.cloud.task.AbstractPollTask;
 import com.sequenceiq.cloudbreak.cloud.template.ResourceChecker;
 
 @Component(PollResourceTask.NAME)
